Refactor Friendbuy browser-destination code to a shared directory.
This will allow some code to be shared between the Friendbuy
browser-destination and destination-action destinations.
