+++ packages/browser-destinations/src/destinations/friendbuy/index.ts
@@ -5,11 +5,13 @@ import { defaultValues } from '@segment/actions-core'
 
 import type { Settings } from './generated-types'
 import type { FriendbuyAPI } from './types'
-import trackCustomer, { trackCustomerDefaultSubscription, trackCustomerFields } from './trackCustomer'
-import trackPurchase, { trackPurchaseDefaultSubscription, trackPurchaseFields } from './trackPurchase'
+import trackCustomer, { trackCustomerDefaultSubscription } from './trackCustomer'
+import trackPurchase, { trackPurchaseDefaultSubscription } from './trackPurchase'
 import trackSignUp, { trackSignUpDefaultSubscription, trackSignUpFields } from './trackSignUp'
 import trackPage, { trackPageDefaultSubscription, trackPageFields } from './trackPage'
 import trackCustomEvent from './trackCustomEvent'
+import { trackCustomerFields } from './shared/sharedCustomer'
+import { trackPurchaseFields } from './shared/sharedPurchase'
 
 declare global {
   interface Window {
