+++ packages/browser-destinations/src/destinations/friendbuy/shared/__tests__/util.test.ts
