+++ packages/browser-destinations/src/destinations/friendbuy/shared/commonFields.ts
