+++ packages/browser-destinations/src/destinations/friendbuy/trackPurchase/__tests__/index.test.ts
@@ -1,6 +1,7 @@
 import { Analytics, Context, JSONValue } from '@segment/analytics-next'
 import friendbuyDestination from '../../index'
-import trackPurchaseObject, { trackPurchaseDefaultSubscription, trackPurchaseFields } from '../index'
+import trackPurchaseObject, { trackPurchaseDefaultSubscription } from '../index'
+import { trackPurchaseFields } from '../../shared/sharedPurchase'
 
 import { loadScript } from '../../../../runtime/load-script'
 jest.mock('../../../../runtime/load-script')
@@ -61,10 +62,10 @@ describe('Friendbuy.trackPurchase', () => {
     const expectedProducts = products.map((p) => {
       p = { quantity: 1, ...p }
       if (p.image_url) {
-        p.imageUrl = p.image_url;
-        delete p.image_url;
+        p.imageUrl = p.image_url
+        delete p.image_url
       }
-      return p;
+      return p
     })
     const amount = expectedProducts.reduce((acc, p) => acc + p.price * p.quantity, 0)
 
@@ -109,8 +110,8 @@ describe('Friendbuy.trackPurchase', () => {
         },
         true
       ])
-      expect(window.friendbuyAPI.push.mock.calls[0][0][2].products[1].quantity).toBe(1);
-      expect(window.friendbuyAPI.push.mock.calls[0][0][2].products[2].imageUrl).toBe(products[2].image_url);
+      expect(window.friendbuyAPI.push.mock.calls[0][0][2].products[1].quantity).toBe(1)
+      expect(window.friendbuyAPI.push.mock.calls[0][0][2].products[2].imageUrl).toBe(products[2].image_url)
     }
 
     {
